+++ src/Valuify/Semantics/INamedTypeSymbolExtensions.HasProperties.cs
@@ -1,6 +1,7 @@
 ﻿namespace Valuify.Semantics;
 
 using Microsoft.CodeAnalysis;
+using Valuify.Model;
 
 /// <summary>
 /// Provides extensions relating to <see cref="INamedTypeSymbol"/>.
@@ -13,13 +14,16 @@ internal static partial class INamedTypeSymbolExtensions
     /// <param name="class">
     /// The class to check for compatible properties.
     /// </param>
+    /// <param name="predicate">
+    /// The predicate to apply to the properties.
+    /// </param>
     /// <returns>
     /// <see langword="true"/> if the <paramref name="class"/> defines compatible properties, otherwise <see langword="false"/>.
     /// </returns>
-    public static bool HasProperties(this INamedTypeSymbol @class)
+    public static bool HasProperties(this INamedTypeSymbol @class, Func<Property, bool> predicate)
     {
         return @class
             .GetAllProperties()
-            .Any();
+            .Any(predicate);
     }
 }